Load messages with site config and move IntlProvider to layout as data is needed later for not found page
